Move client test helpers & matchers to client tests directory
